experiments/random_forest/requirements.txt
pandas
numpy
scikit-learn
gitpython
requests
radon
joblib
fastapi
uvicorn[standard]
pydantic